+++ boxes/definitions/centos-postinstall.sh
@@ -25,9 +25,5 @@ chown -R vagrant /home/vagrant/.ssh
 
 yum -y clean all
 
-# Remove traces of mac address from network configuration
-sed -i /HWADDR/d /etc/sysconfig/network-scripts/ifcfg-eth0
-rm /etc/udev/rules.d/70-persistent-net.rules
-
 # Zero out the free space to save space in the final image
 dd if=/dev/zero of=/filler bs=1M || rm /filler